Apply default row style in WorkbookManager
Instead of doing it in the Writer
